[refactor] refactor dependency inversion data, domain module
