+++ core/data/build.gradle.kts
@@ -8,7 +8,6 @@ android {
 
 dependencies {
     implementation(project(":core:datastore"))
-    implementation(project(":core:domain"))
     implementation(project(":core:model"))
 
     implementation(libs.bundles.retrofit)
