+++ core/data/src/main/kotlin/com/captures2024/soongan/core/data/repository/impl/MembersRepositoryImpl.kt
@@ -1,8 +1,8 @@
-package com.captures2024.soongan.core.data.repository
+package com.captures2024.soongan.core.data.repository.impl
 
 import com.captures2024.soongan.core.data.remote.MembersDataSource
+import com.captures2024.soongan.core.data.repository.MembersRepository
 import com.captures2024.soongan.core.datastore.TokenDataSource
-import com.captures2024.soongan.core.domain.repository.MembersRepository
 import com.captures2024.soongan.core.model.dto.UserInfoDto
 import com.captures2024.soongan.core.model.network.SocialSignType
 import com.captures2024.soongan.core.model.dto.ResultConditionDto
