+++ core/domain/build.gradle.kts
@@ -7,5 +7,6 @@ android {
 }
 
 dependencies {
+    implementation(project(":core:data"))
     implementation(project(":core:model"))
 }
