+++ core/domain/src/main/kotlin/com/captures2024/soongan/core/domain/usecase/members/RegisterNicknameUseCase.kt
@@ -1,6 +1,6 @@
 package com.captures2024.soongan.core.domain.usecase.members
 
-import com.captures2024.soongan.core.domain.repository.MembersRepository
+import com.captures2024.soongan.core.data.repository.MembersRepository
 import com.captures2024.soongan.core.domain.runSuspendCatching
 import javax.inject.Inject
 import javax.inject.Singleton
