+++ core/domain/src/main/kotlin/com/captures2024/soongan/core/domain/usecase/token/SetUUIDUseCase.kt
@@ -1,6 +1,6 @@
 package com.captures2024.soongan.core.domain.usecase.token
 
-import com.captures2024.soongan.core.domain.repository.TokenRepository
+import com.captures2024.soongan.core.data.repository.TokenRepository
 import com.captures2024.soongan.core.domain.runSuspendCatching
 import javax.inject.Inject
 import javax.inject.Singleton
